requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
yt-dlp==2024.10.22
